move CmdGenerateDocs to docs tools
