+++ pulsar-broker/src/main/java/org/apache/pulsar/PulsarTransactionCoordinatorMetadataSetup.java
@@ -23,11 +23,11 @@ import com.beust.jcommander.Parameter;
 import org.apache.pulsar.broker.resources.PulsarResources;
 import org.apache.pulsar.common.naming.NamespaceName;
 import org.apache.pulsar.common.naming.SystemTopicNames;
-import org.apache.pulsar.common.util.CmdGenerateDocs;
+import org.apache.pulsar.docs.tools.CmdGenerateDocs;
 import org.apache.pulsar.metadata.api.extended.MetadataStoreExtended;
 
 /**
- * Setup the transaction coordinator metadata for a cluster, the setup will create pulsar/system namespace and create
+ * Set up the transaction coordinator metadata for a cluster, the setup will create pulsar/system namespace and create
  * partitioned topic for transaction coordinator assign.
  */
 public class PulsarTransactionCoordinatorMetadataSetup {
