+++ pulsar-broker/src/test/java/org/apache/pulsar/PulsarBrokerStarterTest.java
@@ -38,7 +38,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import org.apache.pulsar.broker.ServiceConfiguration;
-import org.apache.pulsar.common.util.CmdGenerateDocs;
+import org.apache.pulsar.docs.tools.CmdGenerateDocs;
 import org.testng.annotations.Test;
 
 @Test(groups = "broker")
